add common_dev to delineate common testing code from common contract code. Ink cannot handle the two together in one crate
