+++ contracts/common_dev/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "captcha"
-version = "4.0.0"
+name = "common_dev"
+version = "0.3.0"
 authors = ["Chris Taylor chris@prosopo.io", "George Oastler george@prosopo.io", "Vincenzo Ferrara", "Siniša Čanak"]
 edition = "2021"
 
@@ -8,39 +8,14 @@ edition = "2021"
 ink = { git = "https://github.com/prosopo/ink", rev="9ca19c462d5f5d1572ddce595d2698207241fd4a", default-features = false }
 scale = { package = "parity-scale-codec", version = "3", default-features = false, features = ["derive"] }
 scale-info = { version = "2.3", default-features = false, features = ["derive"], optional = true }
-hex = "0.4.3"
-
-[dependencies.inject_self_macro]
-version = "0.3.0"
-path = "../../crates/inject_self_macro"
-
-[dependencies.named_functions_macro]
-version = "0.3.0"
-path = "../../crates/named_functions_macro"
-
-[dependencies.disable_macro]
-version = "0.3.0"
-path = "../../crates/disable_macro"
-
-[dependencies.common]
-version = "0.3.0"
-path = "../../contracts/common"
-default-features = false
-features = ["ink-as-dependency"] 
-
-[dev-dependencies.common_dev]
-version = "0.3.0"
-path = "../../contracts/common_dev"
-default-features = false
-features = ["ink-as-dependency"] 
 
 [lib]
-name = "captcha"
+name = "common_dev"
 path = "src/lib.rs"
 crate-type = [
     # Used for normal contract Wasm blobs.
     "cdylib",
-    "rlib"
+    "rlib",
 ]
 
 # Needed until https://github.com/paritytech/ink/issues/364 is resolved.
@@ -49,5 +24,9 @@ overflow-checks = false
 
 [features]
 default = ["std"]
-std = ["ink/std", "scale/std", "scale-info/std"]
+std = [
+    "ink/std",
+    "scale/std",
+    "scale-info/std",
+]
 ink-as-dependency = []
