+++ contracts/proxy/Cargo.toml
@@ -27,6 +27,12 @@ path = "../../contracts/common"
 default-features = false
 features = ["ink-as-dependency"] 
 
+[dev-dependencies.common_dev]
+version = "0.3.0"
+path = "../../contracts/common_dev"
+default-features = false
+features = ["ink-as-dependency"] 
+
 [lib]
 name = "proxy"
 path = "src/lib.rs"
