+++ contracts/proxy/src/lib.rs
@@ -178,7 +178,7 @@ pub mod proxy {
         )
     )]
     mod tests {
-        use common::common::tests::*;
+        use common_dev::common_dev::tests::*;
         use ink;
         use ink::codegen::Env;
         use ink::env::hash::Blake2x256;
